// src/pages/motor/ad-driving-licence-details/aliases.generated.ts
// src/pages/motor/ad-driving-licence-details/aliases.generated.ts
// AUTO-GENERATED by src/scanner/elements-generator
// pageKey: motor.ad-driving-licence-details
// scannedAt: 2026-03-09T13:08:06.096Z

import type { ElementKey } from "./elements";

// Page metadata (used for waitForURL / assertions)
export const pageMeta = {
  pageKey: "motor.ad-driving-licence-details",
  urlPath: "/journey/show/product/AnnualMotorInsurance/process/nb/additionalDriver1DrivingLicenceScanPage/69aec1fabe84b34399895bf0",
  urlRe: /\/journey\/show\/product\/AnnualMotorInsurance\/process\/nb\/additionalDriver1DrivingLicenceScanPage\/[a-z0-9]+/i,
  title: "Driving licence details for additional driver",
  titleRe: new RegExp("Driving licence details for additional driver", "i"),
} as const;

// Default alias mapping: alias key == element key
export const aliasesGenerated = {
  addFrontOfLicence: "addFrontOfLicence" as ElementKey,
  addFrontOfLicence2: "addFrontOfLicence2" as ElementKey,
  additionaldriver1partialdrivinglicencenumberquestionmonthpart: "additionaldriver1partialdrivinglicencenumberquestionmonthpart" as ElementKey,
  additionaldriver1partialdrivinglicencenumberquestionrandompart: "additionaldriver1partialdrivinglicencenumberquestionrandompart" as ElementKey,
  addresslookupwidgetAddresslookupquestioninputBuilding: "addresslookupwidgetAddresslookupquestioninputBuilding" as ElementKey,
  addresslookupwidgetAddresslookupquestioninputPostcode: "addresslookupwidgetAddresslookupquestioninputPostcode" as ElementKey,
  addressOptional: "addressOptional" as ElementKey,
  back: "back" as ElementKey,
  day: "day" as ElementKey,
  euFull: "euFull" as ElementKey,
  euProvisional: "euProvisional" as ElementKey,
  findAddress: "findAddress" as ElementKey,
  firstName: "firstName" as ElementKey,
  howLongHaveAdditionalDriverHeldLicenceFor: "howLongHaveAdditionalDriverHeldLicenceFor" as ElementKey,
  lastName: "lastName" as ElementKey,
  linkToAllianzHomePage: "linkToAllianzHomePage" as ElementKey,
  month: "month" as ElementKey,
  month2: "month2" as ElementKey,
  month3: "month3" as ElementKey,
  month4: "month4" as ElementKey,
  next: "next" as ElementKey,
  no: "no" as ElementKey,
  no2: "no2" as ElementKey,
  no3: "no3" as ElementKey,
  no4: "no4" as ElementKey,
  no5: "no5" as ElementKey,
  other: "other" as ElementKey,
  reactSelect13Input: "reactSelect13Input" as ElementKey,
  reactSelect14Input: "reactSelect14Input" as ElementKey,
  reactSelect15Input: "reactSelect15Input" as ElementKey,
  removeConviction: "removeConviction" as ElementKey,
  removeConviction2: "removeConviction2" as ElementKey,
  removeConviction3: "removeConviction3" as ElementKey,
  title: "title" as ElementKey,
  ukFull: "ukFull" as ElementKey,
  ukFullAutomaticOnly: "ukFullAutomaticOnly" as ElementKey,
  ukProvisional: "ukProvisional" as ElementKey,
  year: "year" as ElementKey,
  year2: "year2" as ElementKey,
  year3: "year3" as ElementKey,
  year4: "year4" as ElementKey,
  yes: "yes" as ElementKey,
  yes2: "yes2" as ElementKey,
  yes3: "yes3" as ElementKey,
  yes4: "yes4" as ElementKey,
  yes5: "yes5" as ElementKey,
} as const;

export type AliasGeneratedKey = keyof typeof aliasesGenerated;